+++ Changelog.md
@@ -1,3 +1,14 @@
+## 4.3.0 - UNRELEASED
+
+### Added
+
+-   Offset configuration feature that lets you define a pre-trigger data length
+    to be recorded before the trigger.
+-   Support for analog trigger activation based on either the rising or falling
+    edge.
+-   Added Digital Triggers and configurable trigger conditions based on digital
+    channel states.
+
 ## 4.2.2 - 2025-05-22
 
 ### Changed